Ultimando detalles
Posible versión final. Falta una última revisión.

+++ src/main/java/trayecto/Trayecto.java
@@ -22,7 +22,7 @@ public abstract class Trayecto {
         this.tramos = tramos;
     }
 
-    public float calcularDemora(){
+    public double calcularDemora(){
         return this.modoDeViaje.demoraSegunModo(this);
     }
     public abstract void cambiarEstado(Transeunte transeunte);
